pyproject.toml
[tool.poetry]
name = "voting_back"
version = "0.0.0"
description = "Basic voting app"
authors = ["Jess Monter"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.6"
python-dotenv = "^0.15.0"
mysqlclient = "^2.0.3"
django-allauth = "^0.44.0"
django-crispy-forms = "^1.11.0"
djangorestframework = "^3.12.2"
django-rest-auth = "^0.9.5"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pre-commit = "^2.10.1"
flake8-isort = "^4.0.0"
flake8-docstrings = "^1.5.0"
pylint = "^2.6.0"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
